Rename package from ap_fits_headers to ap_preserve_header
- Rename package directory to match repository name
- Update pyproject.toml package name and entry points
- Update Makefile targets to use new package name
- Update MANIFEST.in for new package directory
- Update README.md with new package name and install URLs
- Update coverage workflow to use new package name
- Update all Python imports and test mocking paths
