--- ap_fits_headers/__init__.py
@@ -1,7 +1,0 @@
-"""
-Generated By: Cursor (Claude Sonnet 4.5)
-
-ap-fits-headers: Preserve FITS headers from file path key-value pairs.
-"""
-
-__version__ = "0.1.0"
