+++ ap_preserve_header/__init__.py
@@ -1,0 +1,7 @@
+"""
+Generated By: Cursor (Claude Sonnet 4.5)
+
+ap-preserve-header: Preserve FITS headers from file path key-value pairs.
+"""
+
+__version__ = "0.1.0"
