+++ ap_preserve_header/__main__.py
@@ -1,7 +1,7 @@
 # Generated By: Claude Code (Claude Sonnet 4.5)
 """Entry point for python -m invocation."""
 
-from ap_fits_headers.preserve_headers import main
+from ap_preserve_header.preserve_headers import main
 
 if __name__ == "__main__":
     main()
