+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "ap-fits-headers"
+name = "ap-preserve-header"
 version = "0.1.0"
 description = "Preserve FITS headers from file path key-value pairs into .fits and .xisf files"
 readme = "README.md"
@@ -32,7 +32,7 @@ dependencies = [
 ]
 
 [project.scripts]
-ap-fits-headers = "ap_fits_headers.preserve_headers:main"
+ap-preserve-header = "ap_preserve_header.preserve_headers:main"
 
 [project.optional-dependencies]
 dev = [
@@ -44,7 +44,7 @@ dev = [
 ]
 
 [tool.setuptools]
-packages = ["ap_fits_headers"]
+packages = ["ap_preserve_header"]
 
 [tool.setuptools.package-data]
 "*" = ["*.txt", "*.md"]
