+++ tests/test_preserve_headers.py
@@ -1,7 +1,7 @@
 """
 Generated By: Cursor (Claude Sonnet 4.5)
 
-Unit tests for ap_fits_headers.preserve_headers module.
+Unit tests for ap_preserve_header.preserve_headers module.
 """
 
 import os
@@ -11,7 +11,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 from astropy.io import fits
 
-from ap_fits_headers import preserve_headers
+from ap_preserve_header import preserve_headers
 
 
 class TestExtractKeyValuePairs:
@@ -371,7 +371,7 @@ class TestPreserveHeaders:
 class TestMain:
     """Tests for main function."""
 
-    @patch("ap_fits_headers.preserve_headers.preserve_headers")
+    @patch("ap_preserve_header.preserve_headers.preserve_headers")
     def test_main_calls_preserve_headers(self, mock_preserve):
         """Test that main function calls preserve_headers with correct arguments."""
         with patch("sys.argv", ["preserve_headers.py", "/root", "--include", "CAMERA"]):
@@ -383,7 +383,7 @@ class TestMain:
             dryrun=False,
         )
 
-    @patch("ap_fits_headers.preserve_headers.preserve_headers")
+    @patch("ap_preserve_header.preserve_headers.preserve_headers")
     def test_main_with_debug_flag(self, mock_preserve):
         """Test main with --debug flag."""
         with patch(
@@ -398,7 +398,7 @@ class TestMain:
             dryrun=False,
         )
 
-    @patch("ap_fits_headers.preserve_headers.preserve_headers")
+    @patch("ap_preserve_header.preserve_headers.preserve_headers")
     def test_main_with_dryrun_flag(self, mock_preserve):
         """Test main with --dryrun flag."""
         with patch(
@@ -413,7 +413,7 @@ class TestMain:
             dryrun=True,
         )
 
-    @patch("ap_fits_headers.preserve_headers.preserve_headers")
+    @patch("ap_preserve_header.preserve_headers.preserve_headers")
     def test_main_with_include_list(self, mock_preserve):
         """Test main with --include flag."""
         with patch(
@@ -428,7 +428,7 @@ class TestMain:
             dryrun=False,
         )
 
-    @patch("ap_fits_headers.preserve_headers.preserve_headers")
+    @patch("ap_preserve_header.preserve_headers.preserve_headers")
     def test_main_with_all_flags(self, mock_preserve):
         """Test main with all flags."""
         with patch(
